+++ src/main/java/ladder/domain/ladder/Ladder.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.ladder;
 
 public class Ladder {
